Move classes to reasonable packages
