+++ src/main/java/com/github/novotnyr/idea/jwt/rs256/RS256Panel.java
@@ -4,7 +4,7 @@ import com.github.novotnyr.idea.jwt.SignatureContext;
 import com.github.novotnyr.idea.jwt.SignatureContextException;
 import com.github.novotnyr.idea.jwt.ui.secretpanel.SecretPanel;
 import com.github.novotnyr.idea.jwt.ui.secretpanel.SignatureContextChangedListener;
-import com.github.novotnyr.idea.jwt.ui.secretpanel.UnsupportedSignatureContext;
+import com.github.novotnyr.idea.jwt.core.UnsupportedSignatureContext;
 import com.github.novotnyr.idea.jwt.validation.SignatureError;
 import com.intellij.ui.DocumentAdapter;
