+++ src/main/java/com/github/novotnyr/idea/jwt/ui/NewJwtDialog.java
@@ -1,6 +1,6 @@
 package com.github.novotnyr.idea.jwt.ui;
 
-import com.github.novotnyr.idea.jwt.HS256SignatureContext;
+import com.github.novotnyr.idea.jwt.hs256.HS256SignatureContext;
 import com.github.novotnyr.idea.jwt.SignatureContext;
 import com.github.novotnyr.idea.jwt.core.Jwt;
 import com.github.novotnyr.idea.jwt.core.JwtFactory;
